fix: header type mismatch, added fcn to commissioner_app_dummy.cpp

+++ src/app/commissioner_app.hpp
@@ -236,7 +236,7 @@ public:
 
     // Network Diagnostic
     MOCKABLE Error CommandDiagGetQuery(const std::string &aAddr, uint64_t aDiagDataFlags);
-    MOCKABLE Error CommandDiagGetQuery(uint64_t aPeerAloc16, uint64_t aDiagDataFlags);
+    MOCKABLE Error CommandDiagGetQuery(uint16_t aPeerAloc16, uint64_t aDiagDataFlags);
 
     MOCKABLE Error CommandDiagReset(const std::string &aAddr, uint64_t aDiagDataFlags);
